[docker] Ensure logs from the entrypoint are in /var/log
This is useful because it ensures that all of the logs related to
LNT are accessible in the same volume, on /var/log/lnt. This is not
necessary when running the Docker containers locally, however it's
really helpful when running the containers on an EC2 instance where
we only have SSH access. Accessing the Docker logs is more difficult
in these circumstances.