[docs] Overhaul the documentation (#136)
There were multiple mentions of how to run a server, all of them
slightly different. This reorganizes the documentation to have a
single place explaining how to run a server, either locally or
in production via Docker.

This also moves the documentation of the deprecated `nts` test
suite to a less prominent location -- we should actually see if
that can be removed entirely.

Finally, it updates the Quickstart section of "Running tests" to
ensure that the commands actually run on anyone's machine. A
quickstart should not require custom setup.