Drop support for MySQL (#115)
According to Chris Matthews, this had been added exclusively for
the (now defunct) lnt.llvm.org instance, which was running on
Google's internal infrastructure and hence had constraints on
the database software it could use.

Since that is not relevant anymore, simplify the codebase and
the support matrix of LNT by dropping support for MySQL. According
to Chris, LNT was originally always intended to run on a Postgres
database.

This allows us to actually test what we claim to support, making
LNT more robust despite its slightly shrunken support matrix. In
practice, I expect that nobody will be impacted by this change since
nobody is running a MySQL instance.